Added styles for specific elements in navbar
Set the navbar logo height/size. Styled the links in the navbar with style on hover. Used flexbox on the navbar container and set height to occupy the entire space of the navbar. Set the ul to flex to have them sit left to right and added margin for spacing.