Added rules to make the page responsive
Used two different methods to have the layout responsive using flexbox and block styles. This will have the boxes, and the header items go under one another. The height on the header container was set to inherit so that when the page shrinks the header image doesnt overflow and resizes correctly.